clean file organization
