+++ react/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './App.css'
+import './css/App.css'
 // パッケージからインポート
 import yellow from '@material-ui/core/colors/yellow'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
@@ -8,7 +8,7 @@ import { createMuiTheme } from '@material-ui/core/styles'
 import { makeStyles } from '@material-ui/core/styles'
 import { useEffect, useState } from 'react'
 // 他のファイルからインポート
-import Home from './components/Home'
+import Home from './app-components/Home'
 
 const theme = createMuiTheme({
   palette: {
