+++ react/src/app-components/home-components/selection-components/RestaurantInformationDeck.js
