--- react/src/components/ButtonToChangeMode.css
@@ -1,16 +1,0 @@
-.button-to-change-mode{
-    width: 64px;
-    height: 64px;
-    padding: 0;
-    background-color: transparent;
-    border-color: transparent;
-    position: absolute;
-    top: 0;
-    right: 0;
-    cursor: pointer;
-    z-index: 1;
-}
-.button-icon{
-    width: 100%;
-    height: auto;
-}
