--- react/src/components/Buttons.css
@@ -1,6 +1,0 @@
-.Buttons{
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-    text-align: center;
-}
