--- react/src/components/Home.css
@@ -1,50 +1,0 @@
-body {
-  height: 100vh;
-  overflow: hidden;
-  /*height: calc(var(--myvh) * 100);*/
-  background-color: whitesmoke;
-}
-
-/*PCで見るときはサイト下部を空けない*/
-@media screen and (min-width: 767px) {
-  .Home {
-    height: 100vh;
-  }
-}
-/*タブレット・スマホのブラウザで見るときはサイト下部を空ける*/
-@media screen and (max-width: 767px) {
-  .Home {
-    height: 90vh;
-  }
-}
-
-.Home {
-  display: flex;
-  flex-direction: column;
-  max-width: 50vh;
-  margin: 0 auto;
-  position: relative;
-  padding: 0;
-}
-
-.Content-wrapper {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  overflow: hidden;
-  /* IE, Edge 対応 */
-  -ms-overflow-style: none;
-  /* Firefox 対応 */
-  scrollbar-width: none;
-}
-
-.Content {
-  flex: 1;
-  height: calc(10%); /*これがないとflex:1が効かなくなるというバグがある*/
-  padding-bottom: 5px;
-}
-
-/* Chrome, Safari 対応 */
-.Content-wrapper::-webkit-scrollbar {
-  display: none;
-}
