--- react/src/components/RestaurantInformation.css
@@ -1,9 +1,0 @@
-.RestaurantInformation {
-    width: 100%;
-    margin: 3px;
-    position: absolute;
-    -webkit-user-select: none;
-	   -moz-user-select: none;
-	    -ms-user-select: none;
-	        user-select: none;
-}
