--- react/src/components/global.js
@@ -1,4 +1,0 @@
-export let numGlobal = 0
-export const assignNumGlobal = _v => {
-  numGlobal = _v
-};
