+++ react/src/css/Setting.css
@@ -4,7 +4,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  background-image: url("..//..//img/背景＿余白削除.svg");
+  background-image: url("..//img/setting-background.svg");
   /* background-position: center; */
   /* background-size: cover; */
   background-repeat: no-repeat;
