+++ react/src/samples/sampleData2.json
@@ -1,0 +1,63 @@
+[
+    {
+        "Restaurant_id": "2c432768806ddd3011a789f2b5943749c143bb6d",
+        "Name": "ラーメン二郎新代田店",
+        "Address": "東京都新宿区市谷本村町2-19 美術出版アカデミービル1F",
+        "distance_float": 1287.1604419247797,
+        "Distance": "1.3km",
+        "CatchCopy": "鴨調理専門資格「メートルカナルディエ」を世界最年少で取得。メートルカナルディエ認定のシェフだからこそ出来る本物の鴨料理を是非ご堪能ください。",
+        "Price": "4000",
+        "TopRankItem": [],
+        "CassetteOwnerLogoImage": "https://iwiz-olp.c.yimg.jp/c/olp/18/1852ed1e9b6d12c23cd34e79a4be4c71/image/5c/74/logo_restaurantikyu.png",
+        "Category": "ラーメン",
+        "UrlYahooLoco": "https://loco.yahoo.co.jp/place/2c432768806ddd3011a789f2b5943749c143bb6d",
+        "UrlYahooMap": "https://map.yahoo.co.jp/route/walk?from=東京都千代田区紀尾井町1-3&to=東京都新宿区市谷本村町2-19 美術出版アカデミービル1F",
+        "ReviewRating": "4.3",
+        "VotesLike": 9,
+        "VotesAll": 28,
+        "BusinessHour": "10:00～20:00",
+        "Images": [
+            "https://restaurant.img-ikyu.com/rsDatas/rsData114000/r113753/113753ga10000017.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData114000/r113753/113753ga10000018.jpg"
+        ],
+        "RecommendScore": 50
+    },
+    {
+        "Restaurant_id": "f177d0a7e2b62f926f4b939e4146e2bedb4bf021",
+        "Name": "文字数めっちゃ多い時の見た目〜〜！ル・パリジャンル・パリジャンル・パリジャンル・パリジャン",
+        "Address": "東京都新宿区岩戸町19 高野ビル1F",
+        "distance_float": 2410.568750679528,
+        "Distance": "2.4km",
+        "CatchCopy": "ワインおいしいです",
+        "Price": "5000",
+        "LunchPrice": "1800",
+        "DinnerPrice": "5000",
+        "TopRankItem": [],
+        "CassetteOwnerLogoImage": "https://iwiz-olp.c.yimg.jp/c/olp/18/1852ed1e9b6d12c23cd34e79a4be4c71/image/5c/74/logo_restaurantikyu.png",
+        "Category": "フランス料理フランス料理フランス料理フランス料理フランス料理フランス料理フランス料理フランス料理",
+        "UrlYahooLoco": "https://loco.yahoo.co.jp/place/f177d0a7e2b62f926f4b939e4146e2bedb4bf021",
+        "UrlYahooMap": "https://map.yahoo.co.jp/route/walk?from=東京都新宿区新宿&to=東京都新宿区岩戸町19 高野ビル1F",
+        "ReviewRating": "3.7",
+        "VotesLike": 12,
+        "VotesAll": 28,
+        "BusinessHour": "昼11:00~15:00(L.O 14:30) 夜17:00～21:00(L.O 20:30) \n 緊急事態宣言中は昼11:00~15:00(L.O 14:30) 夜17:00～20:00(L.O 19:30)",
+        "Genre": [
+            {
+                "Code": "0102005",
+                "Name": "フランス料理（フレンチ）"
+            }
+        ],
+        "NumberOfParticipants": "1",
+        "Images": [
+            "https://restaurant.img-ikyu.com/rsDatas/rsData105000/r104804/104804ga10000018.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData105000/r104804/104804ga10000019.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData114000/r113849/113849ga10000020.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData114000/r113849/113849ga10000021.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData102500/r102113/102113ga10000013.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData102500/r102113/102113ga10000014.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData109000/r108955/108955ga10000012.jpg",
+            "https://restaurant.img-ikyu.com/rsDatas/rsData109000/r108955/108955ga10000013.jpg"
+        ],
+        "RecommendScore": 100
+    }
+]
